openikt-server/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Deprecated==1.2.14
Django==4.2.11
djangorestframework==3.15.1
gitdb==4.0.11
GitPython==3.1.42
idna==3.7
psycopg2==2.9.9
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.32.2
sh==2.0.6
smmap==5.0.1
sqlparse==0.5.0
typing-extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
uwsgi==2.0.24
xlsxwriter==3.2.0
